+++ NEWS.md
@@ -11,6 +11,7 @@
 * New Parameter Selection section in NCA tab allowing to select parameters by study type (#795)
 * The App optionally maps end of sample collection (AEFRLT) for excretion rate parameter calculations: ERTLST, ERTMAX. (#745)
 * Option to upload multiple input files, which will be bound together to form a single ADNCA data set (#821)
+* Section `General Exclusions` allowing to perform in-App NCA exclusions (#851)
 
 
 ## Bugs fixed